fix circular dep + missing jsonSchema
